requirements/development.txt
-r base.txt

autopep8==1.4.4
pylint==2.3.1
django-debug-toolbar==2.0
pylint-django